// src/setting/Types.ts
export type Translation = {
  [key: string]: string;
};

export type TranslationsKeys = Record<keyof Translation, string>;
// ...
export type Notification = {
  id: number;
  title: string;
  body: string;
  is_read: boolean;
};